Fixed several bugs related to EWS recurrence. Bug 1: When building recurrence object for a weekly pattern, the variable name is very similiar, but use case is different. Fix: Turnery operator for byWeekNo based on which is used.
Bug 2: Re-getting all events updates the wrong recurrence pattern, resulting in variance of results that is incorrect.
Fix: It was due to which pattern was being updated, therefore, it updated all of them. Found the right pattern to update and added a warning if there is 2 or more of the same IDs.